+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/ResourceHandlerTest.java
@@ -16,8 +16,10 @@
 package io.fabric8.kubernetes.client;
 
 import io.fabric8.kubernetes.client.ResourceHandler.Key;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 public class ResourceHandlerTest {
@@ -25,29 +27,29 @@ public class ResourceHandlerTest {
   @Test
   public void testKeyString() {
     Key k = new Key("deployment", "apps/v1");
-    Assert.assertEquals("Key[kind=`deployment`, apiVersion=`apps/v1`]", k.toString());
+    assertEquals("Key[kind=`deployment`, apiVersion=`apps/v1`]", k.toString());
   }
 
   @Test
   public void testKeyHashCode() {
     Key k1 = new Key("deployment", "apps/v1");
     Key k2 = new Key("deployment", "apps/v1");
-    Assert.assertEquals(k1.hashCode(), k2.hashCode());
+    assertEquals(k1.hashCode(), k2.hashCode());
   }
 
   @Test
   public void testKeyEquals() {
     Key k1 = new Key("deployment", "apps/v1");
     Key k2 = new Key("deployment", "apps/v1");
-    Assert.assertEquals(k1, k2);
+    assertEquals(k1, k2);
 
     k1 = new Key("deployment", "apps/v1");
     k2 = new Key("other", "apps/v1");
-    Assert.assertNotEquals(k1, k2);
+    assertNotEquals(k1, k2);
 
     k1 = new Key("deployment", "apps/v1");
     k2 = new Key("deployment", "other");
-    Assert.assertNotEquals(k1, k2);
+    assertNotEquals(k1, k2);
   }
 
 }