+++ kubernetes-model/kubernetes-model/src/test/java/io/fabric8/kubernetes/api/model/ClusterRoleTest.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.fabric8.kubernetes.api.model.rbac.ClusterRole;
 import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBuilder;
 import io.fabric8.kubernetes.api.model.rbac.PolicyRuleBuilder;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static net.javacrumbs.jsonunit.core.Option.*;
 import static net.javacrumbs.jsonunit.fluent.JsonFluentAssert.assertThatJson;