+++ kubernetes-model/kubernetes-model/src/test/java/io/fabric8/kubernetes/api/model/UnmarshallTest.java
@@ -19,12 +19,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.openshift.api.model.Template;
 import io.fabric8.openshift.api.model.TemplateBuilder;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class UnmarshallTest {
 
@@ -49,7 +49,7 @@ public class UnmarshallTest {
 
         //We just want to make sure that it visits nested objects when deserialization from json is used.
         // The exact number is volatile so we just care about the minimum number of objects (list, pod and service).
-        Assert.assertTrue(integer.intValue() >= 3);
+        assertTrue(integer.intValue() >= 3);
 
 
         Template template = (Template) mapper.readValue(getClass().getResourceAsStream("/simple-template.json"), KubernetesResource.class);
@@ -62,7 +62,7 @@ public class UnmarshallTest {
 
         //We just want to make sure that it visits nested objects when deserialization from json is used.
         // The exact number is volatile so we just care about the minimum number of objects (list, pod and service).
-        Assert.assertTrue(integer.intValue() >= 2);
+        assertTrue(integer.intValue() >= 2);
 
 
         ServiceList serviceList = (ServiceList) mapper.readValue(getClass().getResourceAsStream("/service-list.json"), KubernetesResource.class);
@@ -75,6 +75,6 @@ public class UnmarshallTest {
 
         //We just want to make sure that it visits nested objects when deserialization from json is used.
         // The exact number is volatile so we just care about the minimum number of objects (list, pod and service).
-        Assert.assertTrue(integer.intValue() >= 2);
+        assertTrue(integer.intValue() >= 2);
     }
 }