+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/HttpServerTest.java
@@ -23,8 +23,10 @@ import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
 
+@EnableRuleMigrationSupport
 public class HttpServerTest {
   @Rule
   public KubernetesServer server = new KubernetesServer(false);